simplified tabulator based modellist

+++ server/libs/csManager.js
@@ -136,9 +136,11 @@ exports.getSCS = async (itemid,project) => {
 
 exports.deleteModel = async (itemid, project) => {
     let item = await files.findOne({ "_id": itemid, project:project });
-    res = await fetch(conversionServiceURI + '/api/delete/' + item.storageID, { method: 'put'});
-    await files.deleteOne({ "_id": itemid });
-    _updated();
+    if (item) {
+        res = await fetch(conversionServiceURI + '/api/delete/' + item.storageID, { method: 'put'});
+        await files.deleteOne({ "_id": itemid });
+        _updated();
+    }
 };
 
 exports.getPNG = async (itemid, project) => {
